feat: Implement Hash Set
Use Hash Multiplication function to calculate bucket index.

Use golden ratio as A, A * (1 << 32) = 0x61C88647, this is
pre-computed in Linux Kernel source code, choose golden ratio
as A can effectively reduce the bucket index collision.

On the other hand, use BST to store the collided entries in a bucket,
this is for search-time optimization, because we'll do search in all of
`Add`, `Remove` and `Contain` operations.